% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\docType{package}
\name{datasummary}
\alias{MetaPhOR}
\alias{MetaPhOR-package}
\title{MetaPhOR: Metabolic Pathway Analysis of RNA}
\description{
MetaPhOR was developed to enable users to assess metabolic dysregulation using
transcriptomic-level data (RNA-sequencing and Microarray data) and produce
publication-quality figures. A list of differentially expressed genes (DEGs),
which includes fold change and p value, from DESeq2 or limma, can be used as
input, with sample size for MetaPhOR, and will produce a data frame of scores
for each KEGG pathway. These scores represent the magnitude and direction of
transcriptional change within the pathway, along with estimated p-values.
MetaPhOR then uses these scores to visualize metabolic profiles within and
between samples through a variety of mechanisms, including: bubble plots,
heatmaps, and pathway models.
}
\author{
\strong{Maintainer}: Emily Isenhart \email{emily.isenhart@roswellpark.org}

Authors:
\itemize{
    \item Spencer Rosario
}

}
