% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayList.R
\name{pathwayList}
\alias{pathwayList}
\title{List Available Metabolic rWikiPathways}
\usage{
pathwayList()
}
\value{
pathwayList() returns a list of rWikiPathways for use in CytoPath()
}
\description{
List Available Metabolic rWikiPathways
}
\examples{
pathwayList()
}
