% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoPath.R
\name{cytoPath}
\alias{cytoPath}
\title{Map Differentially Expressed Genes to Dysregulated Pathways}
\usage{
cytoPath(
<<<<<<< HEAD
  pathway,
  DEGpath,
  figpath,
  genename,
  headers = c("log2FoldChange", "padj")
=======
    pathway,
    DEGpath,
    figpath,
    genename,
    headers = c("log2FoldChange", "padj")
>>>>>>> upstream/master
)
}
\arguments{
\item{pathway}{character, the name of the pathway to be visualized}

\item{DEGpath}{character, the path to a DEG file by DESeq2 or limma}

\item{figpath}{character, the path to which the figure will be saved}

\item{genename}{character, column name with HUGO Gene Names in DEG file}

<<<<<<< HEAD
\item{headers}{character vector of length 2 in the form c(log fold change
=======
\item{headers}{character vector of length2 in the form c(log fold change
>>>>>>> upstream/master
col name, adjusted p value col name)}
}
\value{
cytoPath() Returns a Cytoscape figure of DEG data on rWikiPathways
}
\description{
<<<<<<< HEAD
requires the package RCy3 and a local instance of Cytoscape
}
\examples{
\donttest{
cytoPath(pathway = "Tryptophan Metabolism",
         DEGpath = system.file("extdata/BRCA_DEGS.csv", package = "MetaPhOR"),
         figpath = file.path(tempdir(), "example_map"),
         genename = "X",
         headers = c("logFC", "adj.P.Val"))
}
=======
Map Differentially Expressed Genes to Dysregulated Pathways
}
\examples{
#open Cytoscape locally before running
cytoPath("Tryptophan Metabolism",
system.file("extdata/BRCA_DEGS.csv", package = "MetaPhOR"),
file.path(tempdir(), "example_map"), "X", headers = c("logFC", "adj.P.Val"))
>>>>>>> upstream/master
}
